tweaks to I3EventHeader has unsigned ints for eventid, etc.
scripts that demonstrate

monolith, etc =>  muxer => firstguess => muon-llh => I3Writer (to .i3 file)
I3Reader (.i3 file) => muxer => muon-llh => I3Writer (different .i3 file)












git-svn-id: http://code.icecube.wisc.edu/svn/projects/dataclasses/branches/dc2@15102 16731396-06f5-0310-8873-f7f720988828

+++ private/dataclasses/physics/I3EventHeader.cxx
@@ -2,8 +2,8 @@
 #include <dataclasses/physics/I3EventHeader.h>
 
 I3EventHeader::I3EventHeader() :
-  runID_(std::numeric_limits<uint64_t>::max()),
-  eventID_(std::numeric_limits<uint64_t>::max())
+  runID_(std::numeric_limits<unsigned int>::max()),
+  eventID_(std::numeric_limits<unsigned int>::max())
 {
 }
 
@@ -22,6 +22,4 @@ I3EventHeader::serialize(Archive& ar, unsigned version)
   ar & make_nvp("EndTime",endTime_);
 }
   
-  
-
 I3_SERIALIZABLE(I3EventHeader);
